Non function script for conzono exact redundancy removal
Like mentioned in the summary it is a non-function script of the conzono exact redundancy removal. One thing extra in this is that instead of using gurobi and solving linear program to check feasibility of generators, we are using algorithm 1 from j scott's paper to see if bounds E and R intersect and if they don't then redundancy has been detected.